Added rotated pictures & Edited
Hi, everyone! I added the rotated pictures, along with the pictures overall. I then edited it a bit to no longer be incomplete, but I may be wrong and/or have a different idea with it than was originally intended, so please feel free to fix my changes with edits of your own.